+++ pycompadre/pycompadre.cpp
@@ -337,7 +337,7 @@ public:
         // get polynomial coefficient size
         const int NP = gmls_object->getPolynomialCoefficientsSize();
         // get number of target sites
-        const int NT = gmls_object->getNeighborListAccessor()->getNumberOfTargets();
+        const int NT = gmls_object->getNeighborLists()->getNumberOfTargets();
 
         Compadre::Evaluator gmls_evaluator(gmls_object);
         auto polynomial_coefficients = gmls_evaluator.applyFullPolynomialCoefficientsBasisToDataAllComponents<double**, Kokkos::HostSpace>
