+++ src/Compadre_GMLS.cpp
@@ -37,7 +37,7 @@ void GMLS::generatePolynomialCoefficients(const int number_of_batches, const boo
     Kokkos::deep_copy(_operations, _host_operations);
 
     // check that if any target sites added, that neighbors_lists has equal rows
-    compadre_assert_release((_neighbor_list_accessor.getNumberOfTargets()==_target_coordinates.extent(0)) 
+    compadre_assert_release((_neighbor_lists.getNumberOfTargets()==_target_coordinates.extent(0)) 
             && "Neighbor lists not set in GMLS class before calling generatePolynomialCoefficients.");
 
     // check that if any target sites are greater than zero (could be zero), then there are more than zero source sites
@@ -61,7 +61,7 @@ void GMLS::generatePolynomialCoefficients(const int number_of_batches, const boo
 
     // initialize all alpha values to be used for taking the dot product with data to get a reconstruction 
     try {
-        int total_neighbors = _neighbor_list_accessor.getTotalNeighborsOverAllListsHost();
+        int total_neighbors = _neighbor_lists.getTotalNeighborsOverAllListsHost();
         int total_added_alphas = _target_coordinates.extent(0)*_added_alpha_size;
         _alphas = decltype(_alphas)("alphas", (total_neighbors + total_added_alphas)*_total_alpha_values*_max_evaluation_sites_per_target);
     } catch(std::exception &e) {
@@ -77,7 +77,7 @@ void GMLS::generatePolynomialCoefficients(const int number_of_batches, const boo
                 std::pow(2,sro.use_target_site_weights), 
                 (sro.transform_type==DifferentEachTarget 
                         || sro.transform_type==DifferentEachNeighbor) ?
-                    _neighbor_list_accessor.getNumberOfTargets() : 1,
+                    _neighbor_lists.getNumberOfTargets() : 1,
                 (sro.transform_type==DifferentEachNeighbor) ?
                     _max_num_neighbors : 1,
                 (sro.output_rank>0) ?
